refactor(noise): fold the built-in rule catalog into one file (13 -> 8)
The five noise_builtin_*.go category tables were created by a single
file-splitting commit (a4430837 "split internal/ai into focused packages")
and have only ever been edited as a unit — 0e4609a3 touched all five at
once, and the reference graph shows zero edges between them. They are pure
data with no seams, so five files bought nothing but folder size.

Merged verbatim into noise_builtin.go with section banners; the category
functions and their append order in collectBuiltinRuleSpecs are unchanged,
so the emitted rule list is byte-identical (verified by diffing the rule
tables against HEAD).

internal/noise drops from 13 to 8 source files and out of
.folder-size-baseline.json entirely.
